[bootloader] Fix the issue with VBE graphics mode and remove debug code
The problem was the truncation of the video mode in:

	vbe_set_video_mode(u8 mode)

Therefore, to completely fix the problem it was enough to just use u16.
Now, the graphics console works when bpp = 32 (the support for bpp = 24 has
not been implemented yet).

+++ bootloader/vbe.c
@@ -52,7 +52,7 @@ bool vbe_get_mode_info(u16 mode, ModeInfoBlock *mi)
    return true;
 }
 
-bool vbe_set_video_mode(u8 mode)
+bool vbe_set_video_mode(u16 mode)
 {
    u32 eax, ebx, ecx, edx, esi, edi;
